+++ client/src/components/AuthorizedNavbar/authorizednavbar.scss
@@ -1,0 +1,2 @@
+@import "../../assets/styles/globalstyles.scss";
+@import "../Navbar/Navbar.scss";
