--- client/src/components/LoggedinNav/authorizednavbar.scss
@@ -1,60 +1,0 @@
-@import "../../assets/styles/globalstyles.scss";
-
-body {
-    margin: 0;
-    font-family: 'Outfit', sans-serif;
-    background-color: white;
-    color: black;
-  }
-  
-  .navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 10px 20px;
-    background-color: white;
-  }
-  
-  .logo-container {
-    display: flex;
-    align-items: center;
-  }
-  
-  .smallLogo {
-    width: 38px;
-    height: 38px;
-    margin-right: 10px;
-  }
-  
-  .navbar .logo {
-    font-size: 35px;
-    font-weight: bold;
-    // color: purple;
-  }
-  
-  .navbar .nav-links {
-    list-style: none;
-    display: flex;
-  }
-  
-  .navbar .nav-links li {
-    margin-left: 20px;
-  }
-  
-  .navbar .nav-links a {
-    text-decoration: none;
-    color: black;
-    font-size: 23px;
-    font-weight: bold;
-  }
-  
-  .navbar .social-icons a {
-    margin-left: 15px;
-    color: black;
-  }
-  
-  .githubLogo,
-  .linkedIn {
-    width: 38px;
-    height: 38px;
-  }
