+++ client/src/components/Navbar/Navbar.scss
@@ -12,6 +12,7 @@
 .logo-container {
     display: flex;
     align-items: center;
+    // border: 2px solid red;
 }
 
 .smallLogo {
@@ -72,8 +73,7 @@
     color: $color-black;
 }
 
-.githubLogo,
-.linkedIn {
+.githubLogo, .linkedIn {
     width: 2.4rem;
     height: 2.4rem;
 }
